+++ src/multigrid/multigrid_Laplace2.F90
@@ -187,7 +187,7 @@ contains
       type(cg_level_connected_T), pointer, intent(in) :: curl  !< pointer to a level for which we approximate the solution
       integer(kind=4),                     intent(in) :: src   !< index of source in cg%q(:)
       integer(kind=4),                     intent(in) :: soln  !< index of solution in cg%q(:)
-      integer,                             intent(in) :: nsmoo !< number of smoothing repetitions
+      integer(kind=4),                     intent(in) :: nsmoo !< number of smoothing repetitions
 
       integer, parameter :: RED_BLACK = 2 !< the checkerboard requires two sweeps
