+++ src/multigrid/multigrid_Laplace4M.F90
@@ -189,10 +189,10 @@ contains
 
       implicit none
 
-      type(cg_level_connected_T), pointer, intent(in) :: curl !< pointer to a level for which we approximate the solution
-      integer(kind=4),                     intent(in) :: src  !< index of source in cg%q(:)
-      integer(kind=4),                     intent(in) :: soln !< index of solution in cg%q(:)
-      integer,                             intent(in) :: nsmoo !< number of smoothing repetitions
+      type(cg_level_connected_T), pointer, intent(in) :: curl  !< pointer to a level for which we approximate the solution
+      integer(kind=4),                     intent(in) :: src   !< index of source in cg%q(:)
+      integer(kind=4),                     intent(in) :: soln  !< index of solution in cg%q(:)
+      integer(kind=4),                     intent(in) :: nsmoo !< number of smoothing repetitions
 
       integer :: n, b
       real    :: L0, Lx, Ly, Lz, Lxy, Lxz, Lyz, iL0
